fix remaining C API drift, expose the ZED X video settings
- ZEDCameraSettings called sl_set_video_settings / sl_get_video_settings, neither
  of which the C API exports; the entry points are sl_set_camera_settings and
  sl_get_camera_settings, and the getter reports through an out parameter
- sl_get_current_min_max_depth returns an ERROR_CODE, not a float, so
  GetCurrentMixMaxDepth was reading the return value out of the wrong register
- CAMERA_SETTINGS stopped at LED_STATUS, so exposure time, analog and digital
  gain, their auto ranges, exposure compensation, denoising, scene illuminance
  and AE anti-banding were unreachable from Unity
- package version follows ZEDCamera.PluginVersion again

+++ ZEDCamera/Assets/SDK/NativeInterface/ZEDCamera.cs
@@ -651,7 +651,7 @@ public static class NativeWrapper
         private static extern float dllz_get_depth_min_range_value(int cameraID);
 
         [DllImport(nameDll, EntryPoint = "sl_get_current_min_max_depth")]
-        private static extern float dllz_get_current_min_max_depth(int cameraID, ref float min, ref float max);
+        private static extern int dllz_get_current_min_max_depth(int cameraID, ref float min, ref float max);
 
 
         /*
